+++ crates/ecstore/src/bucket/lifecycle/mod.rs
@@ -21,6 +21,7 @@ pub mod evaluator;
 pub mod manual_transition_job;
 mod metadata_boundary;
 pub(crate) use metadata_boundary::{LifecycleExpiryConfigs, get_expiry_configs};
+mod object_handlers_common;
 mod object_lock_boundary;
 pub use self::core as lifecycle;
 mod replication_sink;
